+++ src/Cron/Feed.php
@@ -12,7 +12,7 @@ use Omikron\Factfinder\Model\Config\CommunicationConfig;
 use Omikron\Factfinder\Model\Export\FeedFactory as FeedGeneratorFactory;
 use Omikron\Factfinder\Model\FtpUploader;
 use Omikron\Factfinder\Model\StoreEmulation;
-use Omikron\Factfinder\Model\Stream\CsvFactory;
+use Omikron\Factfinder\Api\StreamInterfaceFactory;
 use Omikron\Factfinder\Service\FeedFileService;
 
 class Feed
@@ -24,30 +24,46 @@ class Feed
     private FeedGeneratorFactory $feedGeneratorFactory;
     private StoreManagerInterface $storeManager;
     private CommunicationConfig $communicationConfig;
-    private CsvFactory $csvFactory;
+    private StreamInterfaceFactory $streamFactory;
     private FtpUploader $ftpUploader;
     private PushImport $pushImport;
+    private FeedFileService $feedFileService;
     private string $feedType;
 
+    /**
+     * @param ScopeConfigInterface   $scopeConfig
+     * @param StoreManagerInterface  $storeManager
+     * @param FeedGeneratorFactory   $feedFactory
+     * @param StoreEmulation         $emulation
+     * @param StreamInterfaceFactory $streamFactory
+     * @param FtpUploader            $ftpUploader
+     * @param CommunicationConfig    $communicationConfig
+     * @param PushImport             $pushImport
+     * @param FeedFileService        $feedFileService
+     * @param string                 $type
+     * @SuppressWarnings(PHPMD.ExcessiveParameterList)
+     */
     public function __construct(
         ScopeConfigInterface $scopeConfig,
         StoreManagerInterface $storeManager,
         FeedGeneratorFactory $feedFactory,
         StoreEmulation $emulation,
-        CsvFactory $csvFactory,
+        StreamInterfaceFactory $streamFactory,
         FtpUploader $ftpUploader,
         CommunicationConfig $communicationConfig,
         PushImport $pushImport,
+        FeedFileService $feedFileService,
         string $type
     ) {
         $this->scopeConfig          = $scopeConfig;
         $this->storeManager         = $storeManager;
         $this->feedGeneratorFactory = $feedFactory;
         $this->storeEmulation       = $emulation;
-        $this->csvFactory           = $csvFactory;
+        $this->streamFactory       = $streamFactory;
         $this->ftpUploader          = $ftpUploader;
         $this->communicationConfig  = $communicationConfig;
         $this->pushImport           = $pushImport;
+        $this->feedFileService      = $feedFileService;
         $this->feedType             = $type;
     }
 
@@ -62,7 +78,7 @@ class Feed
                 $storeId = (int) $store->getId();
                 if ($this->communicationConfig->isChannelEnabled($storeId)) {
                     $filename = (new FeedFileService())->getFeedExportFilename($this->feedType, $this->communicationConfig->getChannel());
-                    $stream   = $this->csvFactory->create(['filename' => "factfinder/{$filename}"]);
+                    $stream   = $this->streamFactory->create(['filename' => "factfinder/{$filename}"]);
                     $this->feedGeneratorFactory->create($this->feedType)->generate($stream);
                     $this->ftpUploader->upload($filename, $stream);
                     if ($this->communicationConfig->isPushImportEnabled($storeId)) {